+++ packages/components/src/tabs/style.module.scss
